urgent notice edit done

+++ app/Models/UrgentNotice/UrgentNoticeDocument.php
@@ -15,23 +15,29 @@ class UrgentNoticeDocument extends Model
 
 	public static function addDocuments($documents, $urgent_notice_id)
     {
+        if(isset($documents) && count($documents)>0) {
+            foreach ($documents as $document) {
+                UrgentNoticeDocument::create([
+                    'urgent_notice_id' => $urgent_notice_id,
+                    'document_id' => $document
+                ]);
+            }     
+        }
         
-        foreach ($documents as $document) {
-            UrgentNoticeDocument::create([
-                'urgent_notice_id' => $urgent_notice_id,
-                'document_id' => $document
-            ]);
-        } 
     }
 
     public static function updateDocuments($request, $id)
     {
         $remove_documents = $request['remove_document'];
-        foreach ($remove_documents as $doc) {
-            UrgentNoticeDocument::where('urgent_notice_id', $id)->where('document_id', $doc)->delete();    
+        if(isset($remove_documents) && count($remove_documents)>0) {
+            foreach ($remove_documents as $doc) {
+                UrgentNoticeDocument::where('urgent_notice_id', $id)->where('document_id', $doc)->delete();    
+            }    
         }
+        
 
         $add_documents = $request['urgentnotice_files'];
+
         UrgentNoticeDocument::addDocuments($add_documents, $id);
         
     }
