+++ app/Models/UrgentNotice/UrgentNoticeFolder.php
@@ -14,20 +14,26 @@ class UrgentNoticeFolder extends Model
 
 	public static function addFolders($folders, $urgent_notice_id)
     {
-        foreach ($folders as $folder) {
-            UrgentNoticeFolder::create([
-                'urgent_notice_id' => $urgent_notice_id,
-                'folder_id' => $folder
-            ]);
-        } 
+        if (isset($folders) && count($folders)>0) {
+            foreach ($folders as $folder) {
+                UrgentNoticeFolder::create([
+                    'urgent_notice_id' => $urgent_notice_id,
+                    'folder_id' => $folder
+                ]);
+            }     
+        }
+        
     }
 
     public static function updateFolders($request, $id)
     {
         $remove_folders = $request['remove_folder'];
-        foreach ($remove_folders as $folder) {
-            UrgentNoticeFolder::where('urgent_notice_id', $id)->where('folder_id', $folder)->delete();    
+        if(isset($remove_folders) && count($remove_folders)>0) {
+            foreach ($remove_folders as $folder) {
+               UrgentNoticeFolder::where('urgent_notice_id', $id)->where('folder_id', $folder)->delete();    
+            }    
         }
+        
 
         $add_folders = $request['urgentnotice_folders'];
         UrgentNoticeFolder::addFolders($add_folders, $id);
